add express server. update entry point. rename index csv moved to utils. rework logic to separate cpu and mem logging work #4
